Move PreviewMixin to preview submodule
